Fixed a typo in Gamma function.
S_to_M and M_to_T operators are working for Helmholtz.

Issue to address:

When debugging using 6 sources and 2 targets inside a cube, there are
many empty boxes at the uniform level. However, those boxes are not
marked as empty in the sense that the parent node has pointers to them.

As a result, some work are spawned. These tasks do not affect numerical
results, but can incur unnecessary overhead such as memory allocation.

One situation is M_to_T operator for FMM, which requires some local
scrtach space. When there is no point to process, the operator is still
entered, allocate and deallocate memory and exit.

For large problem sizes, the situation is likely to happen when we deal
with spherical data and large locality counts.