For S_to_M operation, directly work on the expansion buffer inside M, instead of creating a new one. This is fine as the M expects only 1 input here.
Doing so removes @p center from the argument list and make the return
type of S_to_M void.

We still need to set the LCO to decrement its internal counter. This
is done by defining a new entry kStoM in the SetOpCodes enum
class. This code is handled in the same way as kOutEdges inside the
operation handler of the LCO.